doxygen doc, uintptr_t replacements

+++ kernel32/rtl/crc.c
@@ -1,11 +1,11 @@
 #include "crc.h"
 
-uint32_t RtlCrc32(uint32_t initial, uint32_t polynomial, const void *data, uint32_t size) 
+uint32_t RtlCrc32(uint32_t initial, uint32_t polynomial, const void *data, size_t size) 
 {
     const uint8_t *d = data;
     uint32_t crc = initial;
 
-    for(uint32_t i = 0; i < size; ++i)
+    for(size_t i = 0; i < size; ++i)
     {
         crc = crc ^ d[i];
         for(int8_t k = 7; k >= 0; k--)
